packages/modal_services/modal_services/requirements/qwen.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
airportsdata==20260208
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
async-timeout==5.0.1
attrs==25.4.0
blake3==1.0.8
cbor2==5.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
compressed-tensors==0.9.1
cupy-cuda12x==13.6.0
depyf==0.18.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
einops==0.8.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.7
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastrlock==0.8.3
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
gguf==0.10.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mistral_common==1.9.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==2.20.0
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
packaging==26.0
partial-json-parser==0.2.1.1.post7
pillow==12.1.1
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pybind11==3.0.1
pycountry==24.6.1
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.40.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rich-toolkit==0.19.2
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
sentencepiece==0.2.1
sentry-sdk==2.52.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.52.1
sympy==1.13.1
tiktoken==0.12.0
tokenizers==0.21.4
tomli==2.4.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.48.3
triton==3.1.0
typer==0.21.2
typer-slim==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vllm==0.7.3
watchfiles==1.1.1
websockets==16.0
xformers==0.0.28.post3
xgrammar==0.1.11
yarl==1.22.0
zipp==3.23.0
